Se añade la página de ver profile y se mejora el llamado al cackend de otras pantallas

+++ SportsApp/lib/entities/Feedback.dart
@@ -1,7 +1,7 @@
 
 class Feedback{
 
-  int grade;
+  double grade;
   int numberOfPersons;
 
   Feedback(this.grade, this.numberOfPersons);
